Move parallel files to utils
