+++ src/trees/MWNode.h
@@ -7,7 +7,7 @@
 #pragma GCC system_header
 #include <Eigen/Core>
 
-#include "mwparallel/omp.h"
+#include "utils/mrcpp_omp.h"
 #include "macros.h"
 
 #include "trees/MWTree.h"
