+++ src/utils/CMakeLists.txt
@@ -3,6 +3,7 @@ target_sources(mrcpp PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/Plotter.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Printer.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Timer.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/mrcpp_mpi.cpp
 )
 
 list(APPEND utils_h
@@ -10,6 +11,8 @@ list(APPEND utils_h
     ${CMAKE_CURRENT_SOURCE_DIR}/Plotter.h
     ${CMAKE_CURRENT_SOURCE_DIR}/Printer.h
     ${CMAKE_CURRENT_SOURCE_DIR}/Timer.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/mrcpp_mpi.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/mrcpp_omp.h
     )                                                       
                                                                                  
 # Sets install directory for all the headers in the list
